+++ src/dockerfile.ml
@@ -102,6 +102,8 @@ type mount = {
 }
 [@@deriving sexp]
 
+type device = { name : string; required : bool option } [@@deriving sexp]
+
 type healthcheck_options = {
   interval : string option;
   timeout : string option;
@@ -136,11 +138,17 @@ type line =
   | `Comment of string
   | `From of from
   | `Maintainer of string
-  | `Run of mount list * network option * security option * shell_or_exec
+  | `Run of
+    mount list
+    * network option
+    * security option
+    * device option
+    * shell_or_exec
   | `Run_heredoc of
     mount list
     * network option
     * security option
+    * device option
     * (heredoc * string option) list
   | `Cmd of shell_or_exec
   | `Expose of int list
@@ -178,18 +186,22 @@ let crunch l =
   let pack l =
     let rec aux acc = function
       | [] -> acc
-      | `Run (m, n, s, `Shell a) :: `Run (m', n', s', `Shell b) :: tl ->
+      | `Run (m, n, s, d, `Shell a) :: `Run (m', n', s', d', `Shell b) :: tl ->
           if n <> n' then invalid_arg "crunch: at least two networks differ.";
           if s <> s' then invalid_arg "crunch: at least two securities differ.";
-          aux (`Run (merge m m', n, s, `Shells [ a; b ]) :: acc) tl
-      | `Run (m, n, s, `Shells a) :: `Run (m', n', s', `Shell b) :: tl ->
+          if d <> d' then invalid_arg "crunch: devices differ";
+          aux (`Run (merge m m', n, s, d, `Shells [ a; b ]) :: acc) tl
+      | `Run (m, n, s, d, `Shells a) :: `Run (m', n', s', d', `Shell b) :: tl ->
           if n <> n' then invalid_arg "crunch: at least two networks differ.";
           if s <> s' then invalid_arg "crunch: at least two securities differ.";
-          aux (`Run (merge m m', n, s, `Shells (a @ [ b ])) :: acc) tl
-      | `Run (m, n, s, `Shells a) :: `Run (m', n', s', `Shells b) :: tl ->
+          if d <> d' then invalid_arg "crunch: devices differ";
+          aux (`Run (merge m m', n, s, d, `Shells (a @ [ b ])) :: acc) tl
+      | `Run (m, n, s, d, `Shells a) :: `Run (m', n', s', d', `Shells b) :: tl
+        ->
           if n <> n' then invalid_arg "crunch: at least two networks differ.";
           if s <> s' then invalid_arg "crunch: at least two securities differ.";
-          aux (`Run (merge m m', n, s, `Shells (a @ b)) :: acc) tl
+          if d <> d' then invalid_arg "crunch: devices differ";
+          aux (`Run (merge m m', n, s, d, `Shells (a @ b)) :: acc) tl
       | hd :: tl -> aux (hd :: acc) tl
     in
     List.rev (aux [] l)
@@ -340,7 +352,7 @@ let string_of_mount { typ } =
         @ optional_int_octal "mode" mode
         @ optional_int "uid" uid @ optional_int "gid" gid)
 
-let string_of_run' ~escape mounts network security =
+let string_of_run' ~escape mounts network security device =
   let mounts =
     mounts |> List.map string_of_mount
     |> List.map (escape_string ~char_to_escape:' ' ~escape)
@@ -355,15 +367,24 @@ let string_of_run' ~escape mounts network security =
       (function `Insecure -> "insecure" | `Sandbox -> "sandbox")
       security
   in
-  mounts @ network @ security
+  let device =
+    match device with
+    | None -> []
+    | Some { name; required } ->
+        [
+          String.concat ","
+            ([ sprintf "--device=%s" name ] @ optional_flag "required" required);
+        ]
+  in
+  mounts @ network @ security @ device
 
-let string_of_run ~escape mounts network security c =
-  let params = string_of_run' ~escape mounts network security in
+let string_of_run ~escape mounts network security device c =
+  let params = string_of_run' ~escape mounts network security device in
   let run = string_of_shell_or_exec ~escape c in
   String.concat " " (params @ [ run ])
 
-let string_of_run_heredoc ~escape mounts network security c =
-  let params = string_of_run' ~escape mounts network security in
+let string_of_run_heredoc ~escape mounts network security device c =
+  let params = string_of_run' ~escape mounts network security device in
   let escape_cmd = function
     | Some cmd -> " " ^ escape_string ~char_to_escape:'\n' ~escape cmd
     | None -> ""
@@ -395,10 +416,10 @@ let rec string_of_line ~escape (t : line) =
              (match alias with None -> "" | Some a -> " AS " ^ a);
            ])
   | `Maintainer m -> cmd "MAINTAINER" m
-  | `Run (mounts, network, security, c) ->
-      cmd "RUN" (string_of_run ~escape mounts network security c)
-  | `Run_heredoc (mounts, network, security, c) ->
-      cmd "RUN" (string_of_run_heredoc ~escape mounts network security c)
+  | `Run (mounts, network, security, device, c) ->
+      cmd "RUN" (string_of_run ~escape mounts network security device c)
+  | `Run_heredoc (mounts, network, security, device, c) ->
+      cmd "RUN" (string_of_run_heredoc ~escape mounts network security device c)
   | `Cmd c -> cmd "CMD" (string_of_shell_or_exec ~escape c)
   | `Expose pl -> cmd "EXPOSE" (String.concat " " (List.map string_of_int pl))
   | `Arg a -> cmd "ARG" (string_of_arg ~escape a)
@@ -455,18 +476,19 @@ let mount_ssh ?id ?target ?required ?mode ?uid ?gid () =
   let m = { id; target; required; mode; uid; gid } in
   { typ = `Ssh m }
 
+let device ~name ?required () = { name; required }
 let from ?alias ?tag ?platform image = [ `From { image; tag; alias; platform } ]
 let comment fmt = ksprintf (fun c -> [ `Comment c ]) fmt
 let maintainer fmt = ksprintf (fun m -> [ `Maintainer m ]) fmt
 
-let run ?(mounts = []) ?network ?security fmt =
-  ksprintf (fun b -> [ `Run (mounts, network, security, `Shell b) ]) fmt
+let run ?(mounts = []) ?network ?security ?device fmt =
+  ksprintf (fun b -> [ `Run (mounts, network, security, device, `Shell b) ]) fmt
 
-let run_exec ?(mounts = []) ?network ?security cmds : t =
-  [ `Run (mounts, network, security, `Exec cmds) ]
+let run_exec ?(mounts = []) ?network ?security ?device cmds : t =
+  [ `Run (mounts, network, security, device, `Exec cmds) ]
 
-let run_heredoc ?(mounts = []) ?network ?security docs : t =
-  [ `Run_heredoc (mounts, network, security, docs) ]
+let run_heredoc ?(mounts = []) ?network ?security ?device docs : t =
+  [ `Run_heredoc (mounts, network, security, device, docs) ]
 
 let cmd fmt = ksprintf (fun b -> [ `Cmd (`Shell b) ]) fmt
 let cmd_exec cmds : t = [ `Cmd (`Exec cmds) ]