+++ src/dockerfile.mli
@@ -122,11 +122,13 @@ val maintainer : ('a, unit, string, t) format4 -> 'a
 type mount
 type network = [ `Default | `None | `Host ]
 type security = [ `Insecure | `Sandbox ]
+type device
 
 val run :
   ?mounts:mount list ->
   ?network:network ->
   ?security:security ->
+  ?device:device ->
   ('a, unit, string, t) format4 ->
   'a
 (** [run ?mounts ?network ?security fmt] will execute any commands in a new
@@ -145,12 +147,17 @@ val run :
 
     @param security
       Control which security mode the command is run in. Requires BuildKit
-      {{!val:parser_directive}syntax} 1-labs. *)
+      {{!val:parser_directive}syntax} 1-labs.
+
+    @param device
+      Lets builds request CDI devices are available to the build step. See
+      {!val-device}. *)
 
 val run_exec :
   ?mounts:mount list ->
   ?network:network ->
   ?security:security ->
+  ?device:device ->
   string list ->
   t
 (** [run_exec ?mounts ?network ?security args] will execute any commands in a
@@ -169,12 +176,17 @@ val run_exec :
 
     @param security
       Control which security mode the command is run in. Requires BuildKit
-      {{!val:parser_directive}syntax} 1-labs. *)
+      {{!val:parser_directive}syntax} 1-labs.
+
+    @param device
+      Lets builds request CDI devices are available to the build step. See
+      {!val-device}. *)
 
 val run_heredoc :
   ?mounts:mount list ->
   ?network:network ->
   ?security:security ->
+  ?device:device ->
   (heredoc * string option) list ->
   t
 (** [run_heredoc ?mounts ?network ?security docs] will execute any commands in a
@@ -193,7 +205,11 @@ val run_heredoc :
 
     @param security
       Control which security mode the command is run in. Requires BuildKit
-      {{!val:parser_directive}syntax} 1-labs. *)
+      {{!val:parser_directive}syntax} 1-labs.
+
+    @param device
+      Lets builds request CDI devices are available to the build step. See
+      {!val-device}. *)
 
 val mount_bind :
   target:string ->
@@ -312,6 +328,10 @@ val mount_ssh :
 
     @see <https://docs.docker.com/engine/reference/builder/#run---mounttypessh>
       Docker --mount=type=ssh reference *)
+
+val device : name:string -> ?required:bool -> unit -> device
+(** Create a device for [RUN]. Lets builds request CDI devices are available to
+    the build step. *)
 
 val cmd : ('a, unit, string, t) format4 -> 'a
 (** [cmd args] provides defaults for an executing container. These defaults can